refactor the VM to another module
